pyproject.toml
[tool.poetry]
name = "elecciones-salvador"
version = "0.1.0"
description = "Scraping process to download each summary vote totals (JPG or PDF files) published by the Salvador government. There are 16 million votes approximately and the summary vote totals will be used by the independent project of the Guatemalan non-profit organization FundacionHCG.org and FiscalDigital.net"
authors = ["Sebastian Rios Sabogal <sebaxtianrioss@gmail.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
boto3 = "^1.34.33"
python-dotenv = "^1.0.1"
selenium = "^4.17.2"
requests = "^2.31.0"
pyautogui = "^0.9.54"
tk = "^0.1.0"
pandas = "^2.2.0"
loguru = "^0.7.2"
tqdm = "^4.66.2"
numpy = "^1.26.4"
sounddevice = "^0.4.6"
gtts = "^2.5.1"
pygame = "^2.5.2"
pendulum = "^3.0.0"
beautifulsoup4 = "^4.12.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
